requirements.txt
streamlit==1.31.0
openpyxl==3.1.2
pandas==2.2.0
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pdfplumber==0.10.3
Pillow==10.2.0
